# docs/apidoc/apidoc.yaml
openapi: '3.0.0'
info:
  title: Nimbus Workflow API
  version: '0.2.0'
  description: API for managing workflows, tasks, and task runners
servers:
  - url: http://localhost:8080/api/v1
paths:
  /workflows:
    post:
      summary: Create a new workflow
      description: Creates a new workflow with the provided name
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /tasks:
    post:
      summary: Create a new task
      description: Creates a new task with the provided payload
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get all tasks
      description: Retrieves a list of all tasks
      tags:
        - Tasks
      responses:
        '200':
          description: List of tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksListResponse'
        '500':
          $ref: '#/components/responses/500'
  /tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieves a task by its ID
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /tasks/{id}/start:
    post:
      summary: Start a task
      description: Transitions a task from NEW to IN_PROGRESS status. If task runners are assigned, their payloads are dispatched asynchronously.
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to start
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  /tasks/{id}/complete:
    post:
      summary: Complete a task
      description: Transitions a task to COMPLETED status
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to complete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  /tasks/{id}/fail:
    post:
      summary: Fail a task
      description: Transitions a task to FAILED status
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task to fail
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task marked as failed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  /tasks/{taskId}/runners:
    get:
      summary: Get runners assigned to a task
      description: Retrieves all task runners assigned to the specified task
      tags:
        - Task Runners
      parameters:
        - name: taskId
          in: path
          required: true
          description: UUID of the task
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of runners retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunnersListResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /task-runners:
    post:
      summary: Create a task runner
      description: Creates a new task runner configuration
      tags:
        - Task Runners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRunnerRequest'
      responses:
        '201':
          description: Task runner created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunnerResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get all task runners
      description: Retrieves a list of all task runners
      tags:
        - Task Runners
      responses:
        '200':
          description: List of task runners retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunnersListResponse'
        '500':
          $ref: '#/components/responses/500'
  /task-runners/{id}:
    get:
      summary: Get a specific task runner
      description: Retrieves a task runner by its ID
      tags:
        - Task Runners
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task runner
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task runner retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunnerResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /task-runners/{id}/assign/{taskId}:
    post:
      summary: Assign a runner to a task
      description: Creates an assignment between a task runner and a task. When the task is started, its payload will be dispatched to this runner.
      tags:
        - Task Runners
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task runner
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          description: UUID of the task to assign
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task assigned to runner successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Unassign a runner from a task
      description: Removes the assignment between a task runner and a task
      tags:
        - Task Runners
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the task runner
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          description: UUID of the task to unassign
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task unassigned from runner successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'

components:
  schemas:
    CreateWorkflowRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The workflow name
      example:
        name: "My Workflow"
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the workflow
        name:
          type: string
          description: The workflow name
        created_at:
          type: string
          format: date-time
          description: Timestamp when the workflow was created
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        name: "My Workflow"
        created_at: "2025-10-29T10:30:00Z"
    WorkflowResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Workflow'
    CreateTaskRequest:
      type: object
      required:
        - payload
        - workflow_id
      properties:
        payload:
          type: string
          description: The task payload data
        workflow_id:
          type: string
          format: uuid
          description: The ID of the workflow this task belongs to
      example:
        payload: "Sample task payload"
        workflow_id: "123e4567-e89b-12d3-a456-426614174000"
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the task
        payload:
          type: string
          description: The task payload data
        workflow_id:
          type: string
          format: uuid
          description: The ID of the workflow this task belongs to
        status:
          type: string
          enum: [NEW, IN_PROGRESS, COMPLETED, FAILED]
          description: Current status of the task
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        fail_reason:
          type: string
          description: Reason for task failure (present only if status is FAILED)
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        payload: "Sample task payload"
        workflow_id: "123e4567-e89b-12d3-a456-426614174000"
        status: "NEW"
        created_at: "2025-10-29T10:30:00Z"
        fail_reason: ""
    TaskResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Task'
    TasksListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    CreateTaskRunnerRequest:
      type: object
      required:
        - name
        - type
        - config
      properties:
        name:
          type: string
          description: Human-readable name for the runner
        type:
          type: string
          enum: [queue]
          description: The runner backend type
        config:
          type: object
          description: Backend-specific configuration (e.g. queue_url for SQS)
          additionalProperties: true
      example:
        name: "SQS Runner"
        type: "queue"
        config:
          queue_url: "http://localhost:9324/000000000000/nimbus-tasks"
    TaskRunner:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the task runner
        name:
          type: string
          description: Human-readable name for the runner
        type:
          type: string
          enum: [queue]
          description: The runner backend type
        config:
          type: object
          description: Backend-specific configuration
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: Timestamp when the runner was created
      example:
        id: "987e6543-e21b-12d3-a456-426614174000"
        name: "SQS Runner"
        type: "queue"
        config:
          queue_url: "http://localhost:9324/000000000000/nimbus-tasks"
        created_at: "2025-10-29T10:30:00Z"
    TaskRunnerResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskRunner'
    TaskRunnersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskRunner'
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message
      example:
        message: "Task started"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message describing what went wrong

  responses:
    400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Invalid task ID"
    404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Task not found"
    422:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Cannot start task in current state"
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Internal Server Error"